Use new sandbox/_invoke-lambda in HTTP server
