--- src/sandbox/http/runtimes/readme.md
@@ -1,1 +1,0 @@
-WIP! These files wrap a child process call to a Lambda function to setup the request and marshall the response.
